app/requirements.txt
fastapi
uvicorn
boto3
onnxruntime
numpy
pydantic
pytest
requests
azure-storage-blob